+++ src/Rarus/BonusServer/Cards/DTO/TransactionAmount/TransactionAmountCollection.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Rarus\BonusServer\Cards\DTO\TransactionAmount;
+
+/**
+ * Class TransactionAmountCollection
+ *
+ *
+ * @method  attach(TransactionAmount $card, $data = null)
+ * @method  TransactionAmount current()
+ *
+ * @package Rarus\BonusServer\Card\DTO
+ */
+class TransactionAmountCollection extends \SplObjectStorage
+{
+}