+++ src/Rarus/BonusServer/Transactions/DTO/Points/Fabric.php
@@ -4,15 +4,17 @@ declare(strict_types=1);
 namespace Rarus\BonusServer\Transactions\DTO\Points;
 
 use Money\Currency;
-use Money\Money;
+use Money\Currencies\ISOCurrencies;
+use Money\Parser\DecimalMoneyParser;
+
 use Rarus\BonusServer\Cards\DTO\CardId;
 use Rarus\BonusServer\Discounts\DTO\DiscountId;
 use Rarus\BonusServer\Shops\DTO\ShopId;
 use Rarus\BonusServer\Transactions\DTO\CashRegister\CashRegisterId;
 use Rarus\BonusServer\Transactions\DTO\Document\DocumentId;
 use Rarus\BonusServer\Transactions\DTO\Type;
-use Money\Currencies\ISOCurrencies;
-use Money\Parser\DecimalMoneyParser;
+
+use Rarus\BonusServer\Util\DateTimeParser;
 
 /**
  * Class Fabric
@@ -52,5 +54,31 @@ class Fabric
             ->setDiscountId(new DiscountId((string)$arPointTransaction['discount_id']));
 
         return $pointTrx;
+    }
+
+    /**
+     * @param Currency $currency
+     * @param array    $arPoint
+     *
+     * @return Point
+     * @throws \Rarus\BonusServer\Exceptions\ApiClientException
+     */
+    public static function initPointFromServerResponse(Currency $currency, array $arPoint): Point
+    {
+        $moneyParser = new DecimalMoneyParser(new ISOCurrencies());
+
+        $point = new Point();
+        $point
+            ->setSum($moneyParser->parse((string)$arPoint['sum'], $currency->getCode()))
+            ->setDateCreate(DateTimeParser::parseTimestampFromServerResponse((string)$arPoint['date']));
+
+        if ($arPoint['active_from'] !== 0) {
+            $point->setActiveFrom(DateTimeParser::parseTimestampFromServerResponse((string)$arPoint['active_from']));
+        }
+        if ($arPoint['active_to'] !== 0) {
+            $point->setActiveFrom(DateTimeParser::parseTimestampFromServerResponse((string)$arPoint['active_to']));
+        }
+
+        return $point;
     }
 }