+++ src/Rarus/BonusServer/Transactions/DTO/Points/PointCollection.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Rarus\BonusServer\Transactions\DTO\Points;
+
+/**
+ * Class CardCollection
+ *
+ *
+ * @method  attach(Point $point, $data = null)
+ * @method  Point current()
+ *
+ * @package Rarus\BonusServer\Transactions\DTO\Points
+ */
+class PointCollection extends \SplObjectStorage
+{
+}